Beginnings of work on the diary.
Regular events are modelled as DiaryEvents, which have a day of the
week, a time and a duration. The idea is that they will be automatically
scheduled by a regular process, and can be turned off if necessary.

Extra information about a particular week's instance of an event is
modelled by DiaryEventOccurrence.