Add a view of upcoming events.
This doesn't actually show any upcoming events - it's really just a list
of regularly scheduled events as it'd look in a week-to-view diary.
Eventually each event will have a little "Schedule" button next to it,
to add an occurrence of the event for the current week (or future
weeks), and the reverse for removing an occurrence of an event.