No more custom attributes on searches.
The only reason custom attributes were needed was to avoid database
queries for objects, and to customise the way search results were
displayed. All of that is now more elegantly handled by having base
search templates, and then overriding the bits that are needed (the only
reason I needed the custom search index objects was for modifying
titles/subtitles, which are more elegantly handled by blocks in
templates).