Add a method for fetching the range of dates in a series.
I'm not sure if there's a more efficient way to fetch this -
at the moment I'm fetching the dates of all the sermons in
the series. Since there are unlikely to be many sermons in
each series, I suspect this'll be more efficient than two
round trips. It's probably fast enough that it doesn't
matter.