Avoid O(n)ness of page breadcrumbs.
Previously, we had to fetch the ancestors of every page
in our breadcrumb trail, which was redundant (we've already
fetched them for the current page, and the ancestors of our
parent are just our ancestors, minus our parent). Now, we
build up the paths as we go, to avoid the O(n) ancestors
fetch.