Refs #43 - automatically add page children to navigation.
When you add a top-level navigation element with a URL that
points to a non-leaf page, add second-level navigation
elements that point to the page's children.

Still need to keep these up-to-date when adding and removing
pages (not when editing pages, since if the navigation
element is there already it doesn't need changing, and if
it's not, then it must have been manually removed).