Refs #16. Show multi-day events correctly in week-to-view.
This affects both the management interface and the public
interface.

Note that this submission contains a bug (or at least, a
surprising behaviour) - events on any given day are sorted
by their start time, which means that an event that spans
multiple dates and begins on Friday at 3pm, will be shown
in the list of events on Saturyda after an event that
starts on Saturday at 2pm.